Version 1
- The program converts PDF files text, and images and parses them into
PowerPoint Slides which are then saved in a document location of the users choice.
The project provides Students, Teachers, and Businesses e.t.c with a basic Title,
Agenda, and Body(Header at the top) text PPT format. Images and References
also appear in the PPT but further manual input might be required to archieve
a desired look.